houston, we have liftoff (first zexall text output)

+++ trunk/msxtypes.h
@@ -4,7 +4,7 @@
 
 #include "SDL_types.h"
 
-typedef Uint16 nw_byte;         // should be (at least) an 8 bit type
+typedef Uint8 nw_byte;         // should be (at least) an 8 bit type
 typedef Uint16 nw_word;         // should be (at least) an 16 bit type
 
 typedef Uint32 emuTimeType;     // should be (at least) an 32 bit type
